requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
duckduckgo_search==8.0.5
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
logging==0.4.9.6
lxml==6.0.0
openai==1.93.0
primp==0.15.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.7.0
PyPDF2==3.0.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
